requirements.txt
asgiref==3.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
dnspython==2.2.1
mongoengine==0.24.1
pymongo==4.1.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
